requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
matplotlib>=3.5.0
pytest>=7.0.0  # for testing
python-dotenv>=0.19.0  # for configuration management 